refactor: Remove legacy variables
