+++ operate/client/src/App/ProcessInstance/BottomPanel/VariablePanel/Listeners/v2/styled.ts
