+++ operate/client/src/App/ProcessInstance/BottomPanel/VariablePanel/index.tsx
@@ -6,50 +6,90 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {useEffect} from 'react';
+import React, {useEffect, useState} from 'react';
 import {observer} from 'mobx-react';
 
 import {flowNodeSelectionStore} from 'modules/stores/flowNodeSelection';
-import {variablesStore} from 'modules/stores/variables';
 import {TabView} from 'modules/components/TabView';
-import {processInstanceListenersStore} from 'modules/stores/processInstanceListeners';
 import {useProcessInstancePageParams} from '../../useProcessInstancePageParams';
 import {InputOutputMappings} from './InputOutputMappings';
 import {VariablesContent} from './VariablesContent';
-import {Listeners} from './Listeners';
+import {Listeners} from './Listeners/v2';
 import {WarningFilled} from './styled';
+import {useJobs} from 'modules/queries/jobs/useJobs';
+import {useIsRootNodeSelected} from 'modules/hooks/flowNodeSelection';
 
-const VariablePanel = observer(function VariablePanel() {
+// TODO: Remove when listeners tab is fully migrated to v2
+import {processInstanceListenersStore} from 'modules/stores/processInstanceListeners';
+import {Listeners as ListenersLegacy} from './Listeners';
+import {IS_LISTENERS_TAB_V2} from 'modules/feature-flags';
+import type {ListenerEntity} from 'modules/types/operate';
+
+type Props = {
+  setListenerTabVisibility: React.Dispatch<React.SetStateAction<boolean>>;
+};
+
+const VariablePanel: React.FC<Props> = observer(function VariablePanel({
+  setListenerTabVisibility,
+}) {
   const {processInstanceId = ''} = useProcessInstancePageParams();
+  const isRootNodeSelected = useIsRootNodeSelected();
 
   const flowNodeId = flowNodeSelectionStore.state.selection?.flowNodeId;
   const flowNodeInstanceId =
     flowNodeSelectionStore.state.selection?.flowNodeInstanceId;
 
-  const {
-    listenersFailureCount,
-    state,
-    fetchListeners,
-    reset,
-    setListenerTabVisibility,
-  } = processInstanceListenersStore;
-  const {listenerTypeFilter} = state;
+  const [listenerTypeFilter, setListenerTypeFilter] =
+    useState<ListenerEntity['listenerType']>();
 
   const shouldUseFlowNodeId = !flowNodeInstanceId && flowNodeId;
 
+  let jobsFilter = {};
+  if (shouldUseFlowNodeId) {
+    jobsFilter = {
+      processInstanceKey: {$eq: processInstanceId},
+      elementId: {$eq: flowNodeId},
+      ...(listenerTypeFilter && {kind: {$eq: listenerTypeFilter}}),
+    };
+  } else if (flowNodeInstanceId) {
+    jobsFilter = {
+      processInstanceKey: {$eq: processInstanceId},
+      elementId: {$eq: flowNodeInstanceId},
+      ...(listenerTypeFilter && {kind: {$eq: listenerTypeFilter}}),
+    };
+  }
+
+  // TODO: remove when listeners tab is fully migrated to v2
+  const {listenersFailureCount, fetchListeners, reset} =
+    processInstanceListenersStore;
+
+  const {
+    data: jobs,
+    fetchNextPage,
+    fetchPreviousPage,
+    hasNextPage,
+    hasPreviousPage,
+  } = useJobs({
+    payload: {filter: jobsFilter},
+    disabled:
+      !IS_LISTENERS_TAB_V2 || (!shouldUseFlowNodeId && !flowNodeInstanceId),
+    select: (data) => data.pages?.flatMap((page) => page.items),
+  });
+
+  const hasFailedListeners = jobs?.some(({state}) => state === 'FAILED');
+
+  // TODO: remove when listeners tab is fully migrated to v2
   useEffect(() => {
+    if (IS_LISTENERS_TAB_V2) {
+      return;
+    }
     reset();
   }, [flowNodeId, flowNodeInstanceId, reset]);
-
   useEffect(() => {
-    variablesStore.init(processInstanceId);
+    if (IS_LISTENERS_TAB_V2) {
+      return;
+    }
 
-    return () => {
-      variablesStore.reset();
-    };
-  }, [processInstanceId]);
-
-  useEffect(() => {
     if (shouldUseFlowNodeId) {
       fetchListeners({
         fetchType: 'initial',
@@ -88,11 +128,9 @@ const VariablePanel = observer(function VariablePanel() {
           removePadding: true,
           onClick: () => {
             setListenerTabVisibility(false);
-            variablesStore.startPolling(processInstanceId);
-            variablesStore.refreshVariables(processInstanceId);
           },
         },
-        ...(flowNodeSelectionStore.isRootNodeSelected
+        ...(isRootNodeSelected
           ? []
           : [
               {
@@ -101,7 +139,6 @@ const VariablePanel = observer(function VariablePanel() {
                 content: <InputOutputMappings type="Input" />,
                 onClick: () => {
                   setListenerTabVisibility(false);
-                  return variablesStore.stopPolling;
                 },
               },
               {
@@ -110,24 +147,49 @@ const VariablePanel = observer(function VariablePanel() {
                 content: <InputOutputMappings type="Output" />,
                 onClick: () => {
                   setListenerTabVisibility(false);
-                  return variablesStore.stopPolling;
                 },
               },
             ]),
-        {
-          id: 'listeners',
-          testId: 'listeners-tab-button',
-          ...(listenersFailureCount && {
-            labelIcon: <WarningFilled />,
-          }),
-          label: 'Listeners',
-          content: <Listeners />,
-          removePadding: true,
-          onClick: () => {
-            setListenerTabVisibility(true);
-            return variablesStore.stopPolling;
-          },
-        },
+        ...(IS_LISTENERS_TAB_V2
+          ? [
+              {
+                id: 'listeners',
+                testId: 'listeners-tab-button',
+                ...(hasFailedListeners && {
+                  labelIcon: <WarningFilled />,
+                }),
+                label: 'Listeners',
+                content: (
+                  <Listeners
+                    jobs={jobs}
+                    setListenerTypeFilter={setListenerTypeFilter}
+                    fetchNextPage={fetchNextPage}
+                    fetchPreviousPage={fetchPreviousPage}
+                    hasNextPage={hasNextPage}
+                    hasPreviousPage={hasPreviousPage}
+                  />
+                ),
+                removePadding: true,
+                onClick: () => {
+                  setListenerTabVisibility(true);
+                },
+              },
+            ]
+          : [
+              {
+                id: 'listeners',
+                testId: 'listeners-tab-button',
+                ...(listenersFailureCount > 0 && {
+                  labelIcon: <WarningFilled />,
+                }),
+                label: 'Listeners',
+                content: <ListenersLegacy />,
+                removePadding: true,
+                onClick: () => {
+                  setListenerTabVisibility(true);
+                },
+              },
+            ]),
       ]}
       key={`tabview-${flowNodeId}-${flowNodeInstanceId}`}
     />
