+++ operate/client/src/App/ProcessInstance/BottomPanel/VariablePanel/tests/notification.test.tsx
@@ -6,7 +6,7 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {VariablePanel} from './index';
+import {VariablePanel} from '../index';
 import {
   fireEvent,
   render,
