+++ operate/client/src/App/ProcessInstance/BottomPanel/Variables/Footer/CopyVariablesButton.test.tsx
@@ -6,75 +6,197 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {screen} from '@testing-library/react';
 import {CopyVariablesButton} from './CopyVariablesButton';
-import {render} from 'modules/testing-library';
-import {variablesStore} from 'modules/stores/variables';
-import {createVariable} from 'modules/testUtils';
+import {render, screen, waitFor} from 'modules/testing-library';
+import {createVariableV2, createProcessInstance} from 'modules/testUtils';
+import type {QueryVariablesResponseBody} from '@vzeta/camunda-api-zod-schemas';
+import {MemoryRouter, Route, Routes} from 'react-router-dom';
+import {Paths} from 'modules/Routes';
+import {QueryClientProvider} from '@tanstack/react-query';
+import {getMockQueryClient} from 'modules/react-query/mockQueryClient';
+import {mockServer} from 'modules/mock-server/node';
+
+import {writeToClipboard} from './writeToClipboard';
+import {http, HttpResponse} from 'msw';
+
+vi.mock('./writeToClipboard', () => ({
+  writeToClipboard: vi.fn(),
+}));
+const mockWriteToClipboard = vi.mocked(writeToClipboard);
+
+const createWrapper = (processInstanceId = '123456') => {
+  const Wrapper: React.FC<{children?: React.ReactNode}> = ({children}) => (
+    <QueryClientProvider client={getMockQueryClient()}>
+      <MemoryRouter initialEntries={[Paths.processInstance(processInstanceId)]}>
+        <Routes>
+          <Route path={Paths.processInstance()} element={children} />
+        </Routes>
+      </MemoryRouter>
+    </QueryClientProvider>
+  );
+  return Wrapper;
+};
 
 describe('CopyVariableButton', () => {
   it('should be disabled (no variables)', () => {
-    render(<CopyVariablesButton />);
+    const emptyResponse: QueryVariablesResponseBody = {
+      items: [],
+      page: {
+        totalItems: 0,
+      },
+    };
+
+    mockServer.use(
+      http.get(
+        '/v2/process-instances/:processInstanceKey',
+        () => {
+          return HttpResponse.json(
+            createProcessInstance({processInstanceKey: '123456'}),
+          );
+        },
+        {once: true},
+      ),
+      http.post(
+        '/v2/variables/search',
+        () => {
+          return HttpResponse.json(emptyResponse);
+        },
+        {once: true},
+      ),
+    );
+
+    render(<CopyVariablesButton />, {wrapper: createWrapper()});
 
     expect(screen.getByRole('button')).toBeDisabled();
   });
 
   it('should be disabled (too many variables)', () => {
     const variables = [...Array(50)].map((_, i) =>
-      createVariable({name: i.toString()}),
+      createVariableV2({
+        name: i.toString(),
+        value: 'value',
+      }),
     );
 
-    variablesStore.setItems(variables);
+    const response: QueryVariablesResponseBody = {
+      items: variables,
+      page: {
+        totalItems: 100,
+      },
+    };
 
-    render(<CopyVariablesButton />);
+    mockServer.use(
+      http.get(
+        '/v2/process-instances/:processInstanceKey',
+        () => {
+          return HttpResponse.json(
+            createProcessInstance({processInstanceKey: '123456'}),
+          );
+        },
+        {once: true},
+      ),
+      http.post(
+        '/v2/variables/search',
+        () => {
+          return HttpResponse.json(response);
+        },
+        {once: true},
+      ),
+    );
+
+    render(<CopyVariablesButton />, {wrapper: createWrapper()});
 
     expect(screen.getByRole('button')).toBeDisabled();
   });
 
   it('should be disabled (truncated values)', () => {
-    variablesStore.setItems([
-      createVariable({
-        isPreview: true,
-      }),
-    ]);
+    const response: QueryVariablesResponseBody = {
+      items: [createVariableV2({isTruncated: true})],
+      page: {
+        totalItems: 1,
+      },
+    };
 
-    render(<CopyVariablesButton />);
+    mockServer.use(
+      http.get(
+        '/v2/process-instances/:processInstanceKey',
+        () => {
+          return HttpResponse.json(
+            createProcessInstance({processInstanceKey: '123456'}),
+          );
+        },
+        {once: true},
+      ),
+      http.post(
+        '/v2/variables/search',
+        () => {
+          return HttpResponse.json(response);
+        },
+        {once: true},
+      ),
+    );
+
+    render(<CopyVariablesButton />, {wrapper: createWrapper()});
 
     expect(screen.getByRole('button')).toBeDisabled();
   });
 
   it('should copy variables to clipboard', async () => {
-    const writeTextSpy = vi.spyOn(navigator.clipboard, 'writeText');
+    const response: QueryVariablesResponseBody = {
+      items: [
+        createVariableV2({
+          name: 'jsonVariable',
+          value: JSON.stringify({a: 123, b: [1, 2, 3], c: 'text'}),
+        }),
+        createVariableV2({
+          name: 'numberVariable',
+          value: '666',
+        }),
+        createVariableV2({
+          name: 'stringVariable',
+          value: '"text"',
+        }),
+      ],
+      page: {
+        totalItems: 3,
+      },
+    };
 
-    variablesStore.setItems([
-      createVariable({
-        name: 'jsonVariable',
-        value: JSON.stringify({a: 123, b: [1, 2, 3], c: 'text'}),
-      }),
-      createVariable({
-        name: 'numberVariable',
-        value: '666',
-      }),
-      createVariable({
-        name: 'stringVariable',
-        value: '"text"',
-      }),
-    ]);
+    mockServer.use(
+      http.get(
+        '/v2/process-instances/:processInstanceKey',
+        () => {
+          return HttpResponse.json(
+            createProcessInstance({processInstanceKey: '123456'}),
+          );
+        },
+        {once: true},
+      ),
+      http.post(
+        '/v2/variables/search',
+        () => {
+          return HttpResponse.json(response);
+        },
+        {once: true},
+      ),
+    );
 
-    const {user} = render(<CopyVariablesButton />);
+    const {user} = render(<CopyVariablesButton />, {wrapper: createWrapper()});
 
-    expect(
-      screen.getByRole('button', {
-        name: /^copy variables$/i,
-      }),
-    ).toBeEnabled();
+    await waitFor(() => {
+      expect(
+        screen.getByRole('button', {
+          name: 'Copy variables',
+        }),
+      ).toBeEnabled();
+    });
     await user.click(
       screen.getByRole('button', {
-        name: /^copy variables$/i,
+        name: 'Copy variables',
       }),
     );
 
-    expect(writeTextSpy).toHaveBeenCalledWith(
+    expect(mockWriteToClipboard).toHaveBeenCalledWith(
       '{"jsonVariable":{"a":123,"b":[1,2,3],"c":"text"},"numberVariable":666,"stringVariable":"text"}',
     );
   });
