+++ operate/client/src/App/ProcessInstance/BottomPanel/Variables/Footer/index.tsx
@@ -11,7 +11,7 @@ import {PendingVariable} from './PendingVariable';
 import {NewVariable} from './NewVariable';
 import {FooterContainer} from './styled';
 import {AddVariableButton} from './AddVariableButton';
-import {CopyVariablesButton as CopyVariablesButtonV2} from './v2/CopyVariablesButton';
+import {CopyVariablesButton as CopyVariablesButtonV2} from './CopyVariablesButton';
 
 type Props = {
   variant: 'initial' | 'disabled' | 'add-variable' | 'pending-variable';
