+++ operate/client/src/App/ProcessInstance/BottomPanel/Variables/Footer/writeToClipboard.ts
