+++ operate/client/src/App/ProcessInstance/BottomPanel/Variables/tests/mocks.tsx
